create controller for each user
